requirements.txt
nose
nose-html-reporting
enum
pyhamcrest